rename to `PropCombinations`
