--- .storybook/decorators/GeneratedStory/index.tsx
@@ -1,2 +1,0 @@
-export { default, GENERATED_STORY_NAME, PARAM_NAME } from './GeneratedStory';
-export { isGeneratedStory } from './isGeneratedStory';
