+++ .storybook/decorators/PropCombinations/index.tsx
@@ -1,0 +1,2 @@
+export { default, GENERATED_STORY_NAME, PARAM_NAME } from './PropCombinations';
+export { isGeneratedStory } from './isGeneratedStory';
